Apply patch for VFS-269
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/vfs/branches/VFS281@833587 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/commons/vfs/provider/http/HttpFileObject.java
@@ -114,7 +114,7 @@ public class HttpFileObject
             // Assume 0 content-length
             return 0;
         }
-        return Integer.parseInt(header.getValue());
+        return Long.parseLong(header.getValue());
     }
 
     /**
